requirements.txt
beautifulsoup4>=4.12.0
playwright>=1.45.0
selenium>=4.22.0
webdriver-manager>=4.0.2
requests>=2.31.0
deathbycaptcha>=0.1.1
twocaptcha>=0.0.1
cloudscraper>=1.2.71
python-owasp-zap-v2.4>=0.0.20
